+++ packages/core/src/middleware/index.ts
@@ -1,20 +1,31 @@
-import { BunnyRequest, HttpRequestHandlerMethod } from '@bunny-ts/common';
+import {
+  BunnyRequest,
+  GuardsMetadata,
+  HttpRequestHandler,
+  HttpRequestHandlerMethod,
+} from '@bunny-ts/common';
+import { DependencyContainer } from 'core';
 export type Middleware = (
   req: BunnyRequest,
   next: (req: BunnyRequest) => Promise<Response> | Response
 ) => Response | Promise<Response>;
 
 export class MiddlewareFactory {
+  dc: DependencyContainer;
   middlewares: Middleware[] = [];
-
+  constructor(readonly depContainer: DependencyContainer) {
+    this.dc = depContainer;
+  }
   addMiddleware(middleware: Middleware) {
     this.middlewares.push(middleware);
   }
 
-  applyMiddleware(
+  async applyMiddleware(
     req: Request,
-    handler: HttpRequestHandlerMethod
-  ): Response | Promise<Response> {
+    handlerData: HttpRequestHandler
+  ): Promise<Response> {
+    const { handler, controllerToken, handlerName } = handlerData;
+    const controllerInstance = this.dc.resolve(controllerToken);
     let i = 0;
     const next = (request: BunnyRequest): Response | Promise<Response> => {
       if (i === this.middlewares.length) {
@@ -24,6 +35,14 @@ export class MiddlewareFactory {
       i++;
       return middleware(request, next);
     };
-    return next(req as BunnyRequest);
+    const guardsMetadata: GuardsMetadata = Reflect.getMetadata(
+      'guards',
+      controllerInstance as Object,
+      handlerName
+    );
+    console.log(guardsMetadata);
+
+    const response: Response = await next(req as BunnyRequest);
+    return response;
   }
 }